[refactorisation] PEP 621 standard
