+++ poetry.toml
@@ -1,0 +1,5 @@
+[virtualenvs]
+create = true
+in-project = true
+always-copy = false
+system-site-packages = true
