+++ ropgadget/__main__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+
+# Local library imports
+import ropgadget
+
+if __name__ == "__main__":
+    ropgadget.main()
