--- scripts/ROPgadget
@@ -1,12 +1,0 @@
-#!/usr/bin/env python
-## -*- coding: utf-8 -*-
-##
-##  Jonathan Salwan - 2014-05-12 - ROPgadget tool
-##
-##  http://twitter.com/JonathanSalwan
-##  http://shell-storm.org/project/ROPgadget/
-##
-
-import ropgadget
-
-ropgadget.main()
