--- setup.cfg
@@ -1,85 +1,0 @@
-# All configuration for plugins and other utils is defined here.
-# Read more about `setup.cfg`:
-# https://docs.python.org/3/distutils/configfile.html
-
-
-# === Linter configuration ===
-# You can reuse this configuration in your own projects.
-# See: https://wemake-python-stylegui.de/en/latest/pages/usage/integrations/nitpick.html
-
-[flake8]
-# Base flake8 configuration:
-# https://flake8.pycqa.org/en/latest/user/configuration.html
-format = wemake
-show-source = True
-statistics = False
-doctests = True
-
-# Plugins:
-max-complexity = 6
-max-line-length = 100
-
-# Self settings:
-max-imports = 17
-
-# Excluding some directories:
-exclude =
-  .git
-  __pycache__
-  .venv
-  .eggs
-  *.egg
-  dist
-
-# Exclude some checks globally:
-ignore = B001, C901, D100, D101, D102, D103, D104, D107, D205, D400, E221, E241,
-    E251, E266, E271, E272, E722, F403, F405, I004, N802, N803, N806, P101,
-    Q000, W503, W504, WPS110, WPS111, WPS112, WPS122, WPS202, WPS204, WPS210,
-    WPS212, WPS213, WPS214, WPS220, WPS221, WPS222, WPS223, WPS226, WPS229,
-    WPS231, WPS232, WPS301, WPS323, WPS336, WPS337, WPS338, WPS341, WPS347,
-    WPS412, WPS420, WPS421, WPS430, WPS432, WPS433, WPS440, WPS504, WPS514,
-    WPS515, WPS602, WPS605
-
-per-file-ignores =
-    ROPgadget.py: WPS102
-    ropgadget/__init__.py: F401
-    ropgadget/args.py: E501, WPS317, WPS318
-    ropgadget/binary.py: E501
-    ropgadget/core.py: E501, I001, S110, WPS125, WPS503, WPS513, WPS609
-    ropgadget/gadgets.py: C812, E126, E261, E501, E800, WPS318, WPS319, WPS513
-    ropgadget/loaders/__init__.py: F401
-    ropgadget/loaders/elf.py: E126, E203, E222, E501, E701, N801, WPS114,
-                              WPS115, WPS120, WPS318, WPS339
-    ropgadget/loaders/macho.py: E126, E203, E222, E501, N801, WPS114, WPS115,
-                                WPS120, WPS318, WPS339, WPS349
-    ropgadget/loaders/pe.py: E126, E203, E222, E501, N801, WPS114, WPS115,
-                             WPS120, WPS318, WPS339, WPS349
-    ropgadget/loaders/raw.py: E126, E222, E501, WPS318
-    ropgadget/loaders/universal.py: E126, E501, N801, WPS114, WPS115, WPS120,
-                                    WPS318, WPS328, WPS339, WPS519
-    ropgadget/options.py: E501, S110
-    ropgadget/ropchain/__init__.py: F401
-    ropgadget/ropchain/arch/__init__.py: F401
-    ropgadget/ropchain/arch/ropmakerx64.py: E501, S112, W605, WPS211, WPS327,
-                                            WPS503
-    ropgadget/ropchain/arch/ropmakerx86.py: E501, S112, W605, WPS211, WPS327,
-                                            WPS503
-    ropgadget/ropchain/ropmaker.py: E501, N400
-    ropgadget/updateAlert.py: E501, I001, I003, S309, W605, WPS102, WPS301
-    setup.py: E501
-
-[isort]
-# isort configuration:
-# https://github.com/timothycrosley/isort/wiki/isort-Settings
-include_trailing_comma = true
-use_parentheses = true
-# See https://github.com/timothycrosley/isort#multi-line-output-modes
-multi_line_output = 3
-# Is the same as 80 in flake8:
-line_length = 100
-
-
-[darglint]
-# darglint configuration:
-# https://github.com/terrencepreilly/darglint
-strictness = long
